backend/trajectory-manager/requirements.txt
torch
pika
pymysql
cryptography
python-dotenv
numpy
scipy
